requirements.txt
# Discord Bot Requirements
# Install with: pip install -r requirements.txt

# Core Discord library
discord.py>=2.6.4

# HTTP client for API requests
aiohttp>=3.11.16

# Environment variable management
python-dotenv>=0.21.1

# Web search functionality
ddgs>=9.10.0

# Web content extraction
trafilatura>=2.0.0

# PDF processing
pypdf>=6.6.0

# Optional
# Uncomment if needed
# lxml>=6.0.2

